+++ CHANGES.rst
@@ -21,6 +21,9 @@ Changelog
 
 -   Filter Burst-BBH superevents from SNEWS coincidence searches.
 
+-   Don't apply EM_COINC if external event is SubGRBTargeted and vetoed by
+    NOT_GRB.
+
 2.3.6 "Champ" (2024-04-30)
 --------------------------